Custom theme branch updated to latest Release Br 4_2_8

+++ src/com/bsb/hike/utils/Utils.java
@@ -252,7 +252,7 @@ public class Utils
 {
 	private static final String TAG = Utils.class.getSimpleName();
 
-	// Precision points definition for duration logging========================================[[
+    // Precision points definition for duration logging========================================[[
 	public static final class ExecutionDurationLogger
 	{
 		public static final String TAG = ExecutionDurationLogger.class.getSimpleName();
@@ -2662,6 +2662,11 @@ public class Utils
 		return !TextUtils.isEmpty(context.getSharedPreferences(HikeMessengerApp.ACCOUNT_SETTINGS, 0).getString(HikeMessengerApp.NAME_SETTING, null));
 	}
 
+	public static boolean isMsisdnVerified(Context context)
+	{
+		return !TextUtils.isEmpty(context.getSharedPreferences(HikeMessengerApp.ACCOUNT_SETTINGS, 0).getString(HikeMessengerApp.MSISDN_SETTING, null));
+	}
+
 	public static void appStateChanged(Context context)
 	{
 		appStateChanged(context, true, false);
@@ -3046,14 +3051,9 @@ public class Utils
 		{
 			return false;
 		}
-
-		if (file.isDirectory())
+        File listFiles[] = file.listFiles();
+		if (file.isDirectory()  && listFiles != null)
 		{
-			File listFiles[] = file.listFiles();
-			if(listFiles == null)
-			{
-				return false;
-			}
 			for (File f : listFiles)
 			{
 				result = result && deleteFile(f);
@@ -5843,11 +5843,6 @@ public class Utils
 		return createTimelinePostForDPChange(response, true);
 	}
 
-//	public static boolean isDeviceRooted()
-//	{
-//		return RootUtil.isDeviceRooted();
-//	}
-
 	public static boolean isPhotosEditEnabled()
 	{
 		if (!Utils.isUserSignedUp(HikeMessengerApp.getInstance().getApplicationContext(), false))
@@ -7066,10 +7061,18 @@ public class Utils
 		return isAndroidDataStorageDir;
 	}
 
-	/*
+	/**
+	 * Returns a comma seprated string from a given list
+	 */
+	public static String listToString(Collection<String> strings)
+	{
+		return listToString(strings, ",");
+	}
+
+	/**
 	 * Returns a interval added string from a given list
 	 */
-	public static String listToString(List<String> list, String strInterVal)
+	public static String listToString(Collection<String> list, String strInterVal)
 	{
 		if (list == null)
 		{
@@ -8144,14 +8147,15 @@ public class Utils
 
 		try
 		{
-			if (!file.exists() || file.listFiles() == null)
+			File[] listFiles = file.listFiles();
+			if (!file.exists() || listFiles == null)
 			{
 				return 0;
 			}
 
 			int count = 0;
 
-			for (File newFile : file.listFiles())
+			for (File newFile : listFiles)
 			{
 				if (newFile.isDirectory())
 				{
@@ -8379,5 +8383,33 @@ public class Utils
 			}
 		}
 		return false;
+	}
+
+	public static boolean isUnknownUserInfoViewEnabled()
+	{
+		HikeSharedPreferenceUtil prefs = HikeSharedPreferenceUtil.getInstance();
+		return prefs.getData(HikeConstants.ENABLE_UNKNOWN_USER_INFO_IN_CHAT, false);
+	}
+
+	public static String getEmojiByUnicode(int unicode) {
+		return new String(Character.toChars(unicode));
+	}
+
+	public static void recordUpgradeTaskCompletion(String taskKey, long duration)
+	{
+		try
+		{
+			JSONObject json = new JSONObject();
+			json.put(AnalyticsConstants.V2.UNIQUE_KEY, "db_update");
+			json.put(AnalyticsConstants.V2.KINGDOM, "act_hs");
+			json.put(AnalyticsConstants.V2.ORDER, "db_update");
+			json.put(AnalyticsConstants.V2.FAMILY, taskKey);
+			json.put(AnalyticsConstants.V2.GENUS, duration);
+			HAManager.getInstance().recordV2(json);
+		}
+		catch (JSONException e)
+		{
+			e.printStackTrace();
+		}
 	}
 }
